Moved config files to specific config package.
