+++ src/main/java/ro/dev/trellteam/security/config/SpringFoxConfig.java
@@ -1,4 +1,4 @@
-package ro.dev.trellteam.security;
+package ro.dev.trellteam.security.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
